fix: allow fetching source code
This is the first step towards enabling the ability to import modules.
Turns out the window.origin of the sandbox window is always null and
the browser doesn't allow us to make CORS request. The changes here
introduce a fetch RPC to make a fetch request from the sandbox window.